requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-simple-captcha==0.6.0
django-redis==5.4.0
psycopg2-binary==2.9.7

# Code quality tools
flake8==7.0.0
black==24.4.2
isort==5.13.2
mypy==1.10.1

Pillow==10.3.0

openpyxl==3.1.2
python-docx==0.8.11
python-pptx==0.6.21
xlrd==2.0.1
PyMuPDF==1.24.14
PyYAML==6.0.2

oss2==2.18.5
cos-python-sdk-v5==1.9.25
esdk-obs-python==3.25.8
baidubce==0.8.3
qiniu==7.17.0
boto3==1.35.0
webdav==1.1.7

requests==2.31.0
urllib3==2.0.7

pandas==2.1.3
python-dateutil==2.8.2
cryptography==41.0.7

django-debug-toolbar==4.2.0
django-extensions==3.2.3

gunicorn==21.2.0
whitenoise==6.6.0

paho-mqtt==1.6.1
pymodbus==3.5.2

SpeechRecognition==3.10.0
pydub==0.25.1
